pyproject.toml
[project]
name = "ai-agent-debuggy-autodev"
version = "0.1.0"
description = "Build an AI agent in Python"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai==1.12.1",
    "python-dotenv==1.1.0",
]
